Refactoring and UI update.
Uses references instead of pointers for VulkanTypes. It is a lot easier to add new UI images.

+++ VulkanProject/Shaders/shader2d.vert
@@ -3,11 +3,11 @@
 layout(location = 0) in vec3 inPosition;
 layout(location = 1) in vec4 inColor;
 layout(location = 2) in vec2 inTexCoord;
-layout(location = 3) in float inTexLayer;
+layout(location = 3) in uint inTexLayer;
 
 layout(location = 0) out vec4 fragColor;
 layout(location = 1) out vec2 fragTexCoord;
-layout(location = 2) out float fragTexLayer;
+layout(location = 2) out uint fragTexLayer;
 
 void main() {
     gl_Position = vec4(inPosition, 1.0);
